Adding Traits, UU and UB files.
I'm leaving tomorrow and will be gone for maybe a week, so won't be able
to work more on this for a while. Am therefore uploading it here so it
is at least available in case you want to have a look at it. The idea
was to just put this into the three files, and then copy it into the
.xml file when people in the forum have had their say on any errors or
changes.